+++ color_constant.py
@@ -5,4 +5,6 @@ BOLD_BLUE = colorama.Style.BRIGHT + colorama.Fore.BLUE
 BOLD_CYAN = colorama.Style.BRIGHT + colorama.Fore.CYAN
 BOLD_GREEN = colorama.Style.BRIGHT + colorama.Fore.GREEN
 BOLD_MAGENTA = colorama.Style.BRIGHT + colorama.Fore.MAGENTA
+BOLD_RED = colorama.Style.BRIGHT + colorama.Fore.RED
+BOLD_YELLOW = colorama.Style.BRIGHT + colorama.Fore.YELLOW
 RESET = colorama.Fore.RESET