+++ graph.py
@@ -3,8 +3,8 @@ from matplotlib import markers, pyplot
 
 
 def point(point1):
-    if(len(point1) != 2):
-        print('Traceback: Invalid data (2 parameters required)')
+    if not point1:
+        return
     pyplot.figure("Point")
     pyplot.grid()
     pyplot.plot([point1[0]], [point1[1]], marker="o", markersize=5)
@@ -14,8 +14,11 @@ def point(point1):
 
 
 def line(point):
+    if not point:
+        return
     pyplot.figure("Line")
-    pyplot.plot(point[0], point[1], marker='o')
+    pyplot.plot([point[0][0], point[1][0]], [
+                point[0][1], point[1][1]], marker='o')
     pyplot.grid()
     pyplot.xlabel('X axis')
     pyplot.ylabel('Y axis')
@@ -23,9 +26,10 @@ def line(point):
 
 
 def midpoint(point):
+    if not point:
+        return
     pyplot.figure("Midpoint")
-    x = (point[0][0] + point[1][0]) / 2.0
-    y = (point[0][1] + point[1][1]) / 2.0
+    x, y = mid(point)
     pyplot.plot(x, y, marker='o')
     pyplot.plot([point[0][0], point[1][0]], [
                 point[0][1], point[1][1]], marker='o')
@@ -37,6 +41,8 @@ def midpoint(point):
 
 
 def trisect(point):
+    if not point:
+        return
     pyplot.figure("Trisect")
     x = (point[0][0] + point[1][0] * 2) / 3.0
     y = (point[0][1] + point[1][1] * 2) / 3.0
@@ -55,6 +61,8 @@ def trisect(point):
 
 
 def split(point):
+    if not point:
+        return
     pyplot.figure("Section")
     x = (point[0][0] * point[2][1] + point[1][0] *
          point[2][0]) / (point[2][1] + point[2][0])
@@ -71,6 +79,8 @@ def split(point):
 
 
 def distance(point):
+    if not point:
+        return
     pyplot.figure("Length")
     d = sqrt(((point[1][0] - point[0][0]) ** 2) +
              ((point[1][1] - point[1][0]) ** 2))
@@ -83,8 +93,10 @@ def distance(point):
 
 
 def slope(point):
+    if not point:
+        return
     pyplot.figure("Slope")
-    s = (point[1][1] - point[0][1])/(point[1][0] - point[0][0])
+    s = slp(point)
     pyplot.title("Slope: " + str(s))
     pyplot.plot(point[0], point[1], marker='o')
     pyplot.grid()
@@ -92,13 +104,46 @@ def slope(point):
     pyplot.ylabel('Y axis')
     pyplot.show()
 
+
 def perpendicular(point):
+    if not point:
+        return
+    a, b = point
     pyplot.figure("Slope of perpendicular")
-    s = (point[1][1] - point[0][1])/(point[1][0] - point[0][0])
+    x, y = mid((a, b))
+    s = slp(point)
     s = -1 / s
+    pyplot.axline((x, y), slope=s)
     pyplot.title("Slope of perpendicular: " + str(s))
-    pyplot.plot(point[0], point[1], marker='o')
+    pyplot.plot([a[0], b[0]], [a[1], b[1]], marker='o')
     pyplot.grid()
     pyplot.xlabel('X axis')
     pyplot.ylabel('Y axis')
     pyplot.show()
+
+
+def median(point):
+    a, b, c = point
+    m = ((a[0] + b[0] + c[0]) / 3, (a[1] + b[1] + c[1]) / 3)
+    pyplot.figure('Median')
+    pyplot.title(
+        'Point of intersection of medians of the triangle is at \n' + str(m))
+    m = mid((b, c))
+    pyplot.plot([a[0], m[0]], [a[1], m[1]], marker='o')
+    m = mid((a, c))
+    pyplot.plot([b[0], m[0]], [b[1], m[1]], marker='o')
+    m = mid((a, b))
+    pyplot.plot([c[0], m[0]], [c[1], m[1]], marker='o')
+    pyplot.plot([a[0], b[0]], [a[1], b[1]], marker='o', color='blue')
+    pyplot.plot([b[0], c[0]], [b[1], c[1]], marker='o', color='blue')
+    pyplot.plot([c[0], a[0]], [c[1], a[1]], marker='o', color='blue')
+    pyplot.grid()
+    pyplot.show()
+
+
+def mid(point):
+    return (point[0][0] + point[1][0]) / 2.0, (point[0][1] + point[1][1]) / 2.0
+
+
+def slp(point):
+    return (point[1][1] - point[0][1])/(point[1][0] - point[0][0])