+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+# Version 0.1.2
+- New feature: added comments
+- New feature: added support for running commands from files
+- New feature: added command history
+- New feature: added calculation of point of intersection of medians of a triangle. It also shows the three lines of median.
+- New feature: added a shell script file (For Linux) and a batch file (For Windows) to install the python package dependencies
+- New feature: added a batch file (For Windows) to run the program directly using the batch script
+- New feature: added a new details message
+- Minor bugs fixed
+
+# Version 0.1.1
+- Minor bugs fixed
 # Version 0.1.0
 Project published on 27th March 2022